[chore] simplify logic around get_credentials during feature store creation (#1525)
We can simplify the get_credential logic during feature store creation, as we no longer look at the config for credentials. Instead, everything is passed in during the method call to create feature store, and we persist the credentials into mongo.

One longer term benefit of this, is that we can reduce the passing around of get_credentials through a lot of function calls, as the credentials should now always be accessible via the MongoBackedCredentialProvider. This is the first set of changes, as it is the most logically substantive one. Will follow-up with a few more PRs to remove further uses of get_credential if this looks good.