[DEV-1849] Support conditionally updating a feature using a mask derived from other feature(s) (#1582)
The row_index_lineage alignment check is applicable to columns but not
features. Features are joined into a common table and can freely interact with
each other regardless of the lineage.

This overrides the row_index_lineage property for Feature class as a way to
bypass the alignment checks. By doing this, we can support things like
conditionally updating a feature using a mask derived from other features.