[DEV-1996] update compute_target to return observation table instead (#1574) (#1591)
This PR updates the compute_target_table function to return an ObservationTable instead of a TargetTable object. We also make a few changes to start removing the TargetTable API object since that isn't needed anymore.

This change aims to:
- make the SDK a bit simpler, and not introduce another API object
- allow us to reuse the ObservationTable, and be able to pass materialized Targets into the computation of historical features. This will allow us to get a complete set of training data.

Will follow-up to clean up target table a bit more, but getting this out first to get the SDK changes in so that we can release it.