[DEV-2026] Filter EventTable by tile date range directly when possible (#1607)
This adds a filter on EventTable directly based on tile date range when
possible. This is safe to do when there is no lag operations in the applied
transforms. This mainly benefits the cases where a feature involves complex
joins and the existing top level filter cannot be pushed down automatically.