[DEV-1939] api/view: add as_target to view column (#1609)
Add the as_target method into the ViewColumn. This currently doesn't do anything, and only adds the lookup_target node to the graph.

I'll add more unit tests later for other view types, but plan to refactor some of it together with the unit tests for as_feature. As such, just adding a basic one first to dimension_view.