[DEV-1939] lookup_target: add sql generation for lookup target (#1611)
This change refactors some of the code around lookup features, to be reused for lookup targets. In particular, we extract a base lookup aggregator and spec that can be reused between the lookup target, and feature.

We also add some integration tests to test the end to end flow of the lookup target.