cherry pick more target updates to release 0.4 (#1619)
* [chore] notebooks: update notebooks for target (#1608)

* pre-create target

* fix

* rm old

* init

* use target

* updates

* update quick start notebook

* update deep dive notebook

* rm logger

* update notebook

* [DEV-1939] api/view: add as_target to view column (#1609)

Add the as_target method into the ViewColumn. This currently doesn't do anything, and only adds the lookup_target node to the graph.

I'll add more unit tests later for other view types, but plan to refactor some of it together with the unit tests for as_feature. As such, just adding a basic one first to dimension_view.

* [DEV-1939] node/generic: add lookup target node (#1588)

Reuse the lookup node's logic, and created a new lookup target node.

* [DEV-1939] lookup_target: add sql generation for lookup target (#1611)

This change refactors some of the code around lookup features, to be reused for lookup targets. In particular, we extract a base lookup aggregator and spec that can be reused between the lookup target, and feature.

We also add some integration tests to test the end to end flow of the lookup target.

* [chore][workflow] adding DATABRICKS_HTTP_PATH as a secret (#1622)

* adding DATABRICKS_HTTP_PATH as a secret

* adding additional timestamp conversion

* ignore lint

---------

Co-authored-by: Chester <chester@featurebyte.com>