[DEV-2643] Fix observation table sampling not uniform (#1946)
This fixes an issue in observation table creation where the sampling is not
uniform over the full available time range.

Previously in the sampling query, the tablesample clause can potentially return
ordered data and cause some rows to be prematurely truncated by the limit
clause. To prevent that issue, this shuffles the output of tabesample so that
rows are randomly selected before truncation.