[chore] Update feast integration tests to produce non-missing values for features (#2020)
This updates feast integration tests to produce non-missing values for features
by using more appropriate entity values in the request. The get_online_features
function is also updated to allow request timestamp to be provided. It is
provided when used by feature materialise service so that the request timestamp
is consistent throughout.