[DEV-2955] Move table row index handling of materialized tables to backend (#2069)
Previously, the logic of removing row index column when previewing, sampling
and downloading materialized tables is in the SDK client. The downside is that
the column is still visible elsewhere such as in the UI.

To fix that, this PR moves such logic to the backend so that the API routes
return consistent result. This also adds sorting by row index column when
downloading materialised table if such a column is available.