[DEV-2985] Add special handling for features with composite entity (#2093)
This adds special handling for features with composite entity so that they work
with feast which doesn't support entity with multiple join keys. In the offline
store feature tables, serving names of the composite entity are concatenated as
a new column which will be used as the entity's join key.