[DEV-2854] Use stored entity relationships in sql generation (part 1) (#2189)
This updates FeatureExecutionPlanner to accept an optional
EntityRelationshipsContext parameter that provides information about the stored
relationships (passed in through the ParentServingPreparation class). When
provided, sql generation will use the entity relationships stored within
features to perform parent entity lookups that are feature specific.

The new parameter is currently only specified in unit tests, so there is no
functional changes. It will be activated in the next PR which requires changes
to FeatureListModel.