[DEV-2854] Include serving name in internal aggregated column names (#2194)
This updates the internal aggregated column names in the generated sql queries
for feature compute to include serving names. This is a follow up to #2189,
after which joining from the same source may produce different result based on
the join keys used.

Since the sql queries are transient, there is no backward compatibility
concerns.