[DEV-2854] Use stored entity relationships in sql generation (part 2) (#2196)
This updates the FeatureCluster to track entity relationships stored in
features. EntityValidationService will now populate the
entity_relationships_context parameter when handling eligible feature lists so
that sql generation will use frozen relationships.

One additional change is to store the entity column names in
RelationshipInfoModel: this is required because we cannot dynamically retrieve
them from table model as the entities can be untagged.