[chore] Make SnowflakeSession connection initialisation more self-contained (#2228)
This updates the _initialize_connection method of SnowflakeSession to be more
self-contained by setting the required parameters for the session such as
timezone and user role.

The error handling when establishing a connection is also updated to surface
the actual error rather than always assuming that the error was caused by
credentials issue. For example, if the error was due to network issue, it
should surface a more accurate error message.