[DEV-3136] Improve efficiency of construct_tile_gen_sql in TileCache (#2232)
This improves the efficiency of construct_tile_gen_sql when used by TileCache,
specifically in the _get_unique_tile_infos method.

The problem is that GraphInterpreter repeatedly flattens the same query graph
when processing each node, but graph flattening is expensive and we should
avoid doing that.

To fix the issue, this refactors GraphInterpreter to flatten the graph once and
reuse the result when processing different nodes. It also removes some of the
deep copies made by sqlglot when constructing the sql which are not necessary
and affect runtime.