[DEV-3306] Improve precomputed lookup feature tables handling (#2319)
This attempts to improve the handling of precomputed lookup feature tables to
be more resilient to unexpected errors:

1. Split deployment enable and disable handling into different methods. This is
so that when disabling a deployment, there is no way we will try to do things
like re-initializing a missing lookup feature table (e.g on bad state), and
vice versa.

2. In feature materialize service's _get_latest_from_feature_table(), handle
the case where the data warehouse feature table is missing some expected
columns based on our mongo records. This prevents it from failing immediately,
and we have a chance to recover when initialize_new_columns() is called
subsequently on the table.