[DEV-3331] Fix feast get online features error for composite entities (#2335)
This fixes a ValueError that is raised by feast get_online_features method in
some cases. The error is triggered when a feature uses a feature table with
composite entity that undergoes precomputed entity lookup (e.g. state x product
to customer x product, where state is a parent of customer).