[DEV-3232] Make online disable handling more robust to unexpected failures (#2337)
This updates the online disable handling for offline feature store tables to be
more robust to unexpected / previous errors.

Changes:

1. When updating offline store feature table documents, obtain the most up to
date features that should have been offline disabled, instead of just the
current list of features to be disabled. This way it can correct the states if
a previous online disabling action failed unexpectedly.

2. When cleaning up offline store feature tables in the data warehouse, handle
and log unexpected errors to allow the disabling process to continue instead of
being terminated.